# platform-api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hyperfleet Platform API
  description: Platform API for ROSA HCP regional cluster management
  version: 0.0.1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v0
    description: API v0
paths:
...
  /management_clusters:
    post:
      summary: Create a new management cluster
      description: |
        Creates a new management cluster by storing a ManagementCluster custom resource.
        Requires privileged access (admin AWS account).
      operationId: createManagementCluster
      tags:
        - ManagementClusters
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManagementClusterRequest"
      responses:
        "201":
          description: Management cluster created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementCluster"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden - account not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Management cluster already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all management clusters
      description: |
        Returns a list of all management clusters.
        Requires privileged access (admin AWS account).
        Note: pagination parameters are not currently implemented; all clusters
        are returned in a single response.
      operationId: listManagementClusters
      tags:
        - ManagementClusters
      responses:
        "200":
          description: List of management clusters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementClusterList"
        "403":
          description: Forbidden - account not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /management_clusters/{id}:
    get:
      summary: Get a management cluster by ID
      description: |
        Returns a specific management cluster by its ID.
        Requires privileged access (admin AWS account).
      operationId: getManagementCluster
      tags:
        - ManagementClusters
      parameters:
        - name: id
          in: path
          required: true
          description: Management cluster ID
          schema:
            type: string
      responses:
        "200":
          description: Management cluster details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementCluster"
        "403":
          description: Forbidden - account not privileged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Management cluster not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Cluster Management Endpoints
...
components:
  schemas:
    ManagementClusterRequest:
      type: object
      description: Request body for creating a management cluster
      required:
        - id
        - region
        - accountId
      properties:
        id:
          type: string
          description: Unique identifier for the management cluster
        region:
          type: string
          description: AWS region of the management cluster
        accountId:
          type: string
          description: AWS account ID that owns the management cluster
    ManagementCluster:
      type: object
      description: A management cluster registered on hyperfleet-db
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier for the management cluster
        region:
          type: string
          description: AWS region of the management cluster
        accountId:
          type: string
          description: AWS account ID that owns the management cluster
    ManagementClusterList:
      type: object
      description: List of management clusters
      required:
        - kind
        - total
        - items
      properties:
        kind:
          type: string
          description: Resource kind
          example: ManagementClusterList
        total:
          type: integer
          description: Total number of items
        items:
          type: array
          items:
            $ref: "#/components/schemas/ManagementCluster"
    Error:
      type: object
      description: Error response
      required:
        - kind
        - code
        - reason
      properties:
        kind:
          type: string
          description: Resource kind
          example: Error
        code:
          type: string
          description: Error code
        reason:
          type: string
          description: Human-readable error message
...